+++ bootstrap/servicemanager.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/ixocreate
- * @copyright IXOCREATE GmbH
+ * @copyright IXOLIT GmbH
  * @license MIT License
  */
 
@@ -12,11 +12,13 @@ namespace Ixocreate\Schema;
 use Ixocreate\Application\Service\ServiceManagerConfigurator;
 use Ixocreate\Schema\Builder\Builder;
 use Ixocreate\Schema\Element\ElementSubManager;
+use Ixocreate\Schema\Link\LinkManager;
 use Ixocreate\Schema\Type\TypeSubManager;
 
 /** @var ServiceManagerConfigurator $serviceManager */
 $serviceManager->addSubManager(ElementSubManager::class);
 $serviceManager->addSubManager(SchemaSubManager::class);
 $serviceManager->addSubManager(TypeSubManager::class);
+$serviceManager->addSubManager(LinkManager::class);
 
 $serviceManager->addFactory(Builder::class);