+++ tests/bootstrap.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link https://github.com/ixocreate
- * @copyright IXOCREATE GmbH
+ * @copyright IXOLIT GmbH
  * @license MIT License
  */
 
@@ -9,5 +9,8 @@ declare(strict_types=1);
 
 namespace Test\Template;
 
+use DG\BypassFinals;
+
 \chdir(\dirname(__DIR__));
 include 'vendor/autoload.php';
+BypassFinals::enable();