+++ tests/src/Link/LinkBootstrapItemTest.php
@@ -1,0 +1,28 @@
+<?php
+/**
+ * @link https://github.com/ixocreate
+ * @copyright IXOLIT GmbH
+ * @license MIT License
+ */
+
+declare(strict_types=1);
+
+namespace Ixocreate\Test\Schema\Link;
+
+use Ixocreate\Schema\Link\LinkBootstrapItem;
+use Ixocreate\Schema\Link\LinkConfigurator;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \Ixocreate\Schema\Link\LinkBootstrapItem
+ */
+class LinkBootstrapItemTest extends TestCase
+{
+    public function testBootstrapItem()
+    {
+        $bootstrapItem = new LinkBootstrapItem();
+        $this->assertInstanceOf(LinkConfigurator::class, $bootstrapItem->getConfigurator());
+        $this->assertSame('link', $bootstrapItem->getVariableName());
+        $this->assertSame('link.php', $bootstrapItem->getFileName());
+    }
+}