+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/schema-package/compare/0.2.1...develop)
+## [Unreleased](https://github.com/ixocreate/schema-package/compare/0.2.2...develop)
+
+## [v0.2.2 (2019-05-27)](https://github.com/ixocreate/schema-package/compare/0.2.1...0.2.2)
+### Added
+- Link bootstrap item & LinkManager 
+### Changed
+- Use LinkManager for LinkType
 
 ## [v0.2.1 (2019-05-20)](https://github.com/ixocreate/schema-package/compare/0.2.0...0.2.1)
 ### Fixed